+++ CHANGELOG.md
@@ -1,8 +1,22 @@
 # Plume-Util change log
 
-## 1.1.6 (forthcoming)
+## 1.1.7
 
-Add UtilPlume.sleep method
+- New class `UniqueIdMap`
+- New methods:
+   - `UtilPlume.replacePrefix`
+   - `UtilPlume.replaceSuffix`
+
+## 1.1.6
+
+- New interface `UniqueId` for objects that have a unique ID.
+- New methods:
+   - `UtilPlume.sleep`
+   - `UtilPlume.countFormatArguments`
+   - `UtilPlume.prefixLines`
+   - `UtilPlume.prefixLinesExceptFirst`
+   - `UtilPlume.indentLines`
+   - `UtilPlume.indentLinesExceptFirst`
 
 ## 1.1.5